+++ public/sw.js
@@ -1,4 +1,4 @@
-// sw-workbox.js - Workbox Service Worker para Qumran Watch v13.1.63
+// sw-workbox.js - Workbox Service Worker para Qumran Watch v13.1.64
 importScripts('https://storage.googleapis.com/workbox-cdn/releases/7.1.0/workbox-sw.js');
 
 const { precacheAndRoute, cleanupOutdatedCaches } = workbox.precaching;
@@ -7,9 +7,9 @@ const { NetworkFirst, CacheFirst, StaleWhileRevalidate } = workbox.strategies;
 const { ExpirationPlugin } = workbox.expiration;
 const { CacheableResponsePlugin } = workbox.cacheableResponse;
 
-console.log('[SW] Workbox SW v13.1.63 iniciando...');
+console.log('[SW] Workbox SW v13.1.64 iniciando...');
 
-precacheAndRoute([{"revision":"830bb116a513550c5858d60ded660753","url":"privacy.html"},{"revision":"68d7883c577c7706f671727e263e7dc8","url":"manifest.json"},{"revision":"c9572888756e5c887d1b56b6dff80e51","url":"license.html"},{"revision":"f0e88f9daff00f98b77ea29a75e80f56","url":"index.html"},{"revision":"f214d4ac2c7f2e2c94e366ca34c5c92e","url":"icon.png"},{"revision":"06d4ba244eaa8030bcd1735a833ce8c0","url":"src/js/index.js"},{"revision":"8bcd3fef48da3300eb0d1b52913f9ff1","url":"src/js/core.js"},{"revision":"961b051e8832d3d48163b7649f872924","url":"src/css/index.css"},{"revision":"2d8904c9b0cd7cb2929d0bb613047f58","url":"src/css/fonts/david-libre-v17-latin-regular.woff2"},{"revision":"1d9878b23b606fc71d20a4ed5bd2ce1f","url":"src/css/fonts/david-libre-v17-latin-700.woff2"},{"revision":"63126eeda8319b882d8d9320edc8ca5a","url":"src/css/fonts/cinzel-v26-latin-regular.woff2"},{"revision":"82175d611596bd4529ca178ff59fb1f4","url":"src/css/fonts/cinzel-v26-latin-700.woff2"},{"revision":"f3cbad24e40bcc40a28cb3422428f6f3","url":"screenshots/screenshot-mobile.png"},{"revision":"501b5f4b3c05c1493b3159232202824f","url":"screenshots/screenshot-desktop.png"},{"revision":"cfffe1bdc43de67a79dda9bfda56921a","url":"icons/icon-96x96.png"},{"revision":"5a05b85f35be2bccc989a6944fb64c01","url":"icons/icon-72x72.png"},{"revision":"5275b3630c847fb0d6eca560d7003b5e","url":"icons/icon-48x48.png"},{"revision":"fc19164b1b3c0a5ef1cec05215e48f82","url":"icons/icon-32x32.png"},{"revision":"29e75c92dc87a39f004eb34346e7ce24","url":"icons/icon-192x192.png"},{"revision":"ebc09dae95e62cf426329bf34dcfd363","url":"icons/icon-144x144.png"},{"revision":"f2816678e93e4302b103ff71f70e8da2","url":"assets/sun-worker.js"}] || []);
+precacheAndRoute([{"revision":"830bb116a513550c5858d60ded660753","url":"privacy.html"},{"revision":"6944834e07931964ff7a6b8b29ec2a4c","url":"manifest.json"},{"revision":"c9572888756e5c887d1b56b6dff80e51","url":"license.html"},{"revision":"f0e88f9daff00f98b77ea29a75e80f56","url":"index.html"},{"revision":"f214d4ac2c7f2e2c94e366ca34c5c92e","url":"icon.png"},{"revision":"55aaa86f5000830453895cf06fe76d67","url":"src/js/index.js"},{"revision":"8bcd3fef48da3300eb0d1b52913f9ff1","url":"src/js/core.js"},{"revision":"961b051e8832d3d48163b7649f872924","url":"src/css/index.css"},{"revision":"2d8904c9b0cd7cb2929d0bb613047f58","url":"src/css/fonts/david-libre-v17-latin-regular.woff2"},{"revision":"1d9878b23b606fc71d20a4ed5bd2ce1f","url":"src/css/fonts/david-libre-v17-latin-700.woff2"},{"revision":"63126eeda8319b882d8d9320edc8ca5a","url":"src/css/fonts/cinzel-v26-latin-regular.woff2"},{"revision":"82175d611596bd4529ca178ff59fb1f4","url":"src/css/fonts/cinzel-v26-latin-700.woff2"},{"revision":"f3cbad24e40bcc40a28cb3422428f6f3","url":"screenshots/screenshot-mobile.png"},{"revision":"501b5f4b3c05c1493b3159232202824f","url":"screenshots/screenshot-desktop.png"},{"revision":"cfffe1bdc43de67a79dda9bfda56921a","url":"icons/icon-96x96.png"},{"revision":"5a05b85f35be2bccc989a6944fb64c01","url":"icons/icon-72x72.png"},{"revision":"5275b3630c847fb0d6eca560d7003b5e","url":"icons/icon-48x48.png"},{"revision":"fc19164b1b3c0a5ef1cec05215e48f82","url":"icons/icon-32x32.png"},{"revision":"29e75c92dc87a39f004eb34346e7ce24","url":"icons/icon-192x192.png"},{"revision":"ebc09dae95e62cf426329bf34dcfd363","url":"icons/icon-144x144.png"},{"revision":"f2816678e93e4302b103ff71f70e8da2","url":"assets/sun-worker.js"}] || []);
 cleanupOutdatedCaches();
 
 self.addEventListener('message', (event) => {
@@ -176,4 +176,4 @@ async function showDueNotifications() {
     }
 }
 
-console.log('[SW] Workbox SW v13.1.63 listo');
+console.log('[SW] Workbox SW v13.1.64 listo');